Add EngiDB and Kotoba Types Crates for Enhanced Graph Management
- Introduced new crates `engidb` and `kotoba-types` to modularize the project, improving organization and maintainability.
- Updated `Cargo.toml` to include workspace members for the new crates, facilitating better dependency management.
- Enhanced `Cargo.lock` to reflect the addition of new dependencies and versions, ensuring compatibility.
- Refactored existing code to utilize the new crates, improving clarity and structure in the codebase.
- Implemented core functionalities in `engidb` for graph database operations and in `kotoba-types` for unified IR types, laying the groundwork for future features and enhancements.
