+++ crates/engidb/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "engidb"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+kotoba-types = { path = "../kotoba-types" }
+redb = "2.1.0"
+cid = { version = "0.11.1", features = ["serde-codec"] }
+serde_ipld_dagcbor = "0.4.0"
+multihash = "0.19.3"
+blake3 = "1.5.1"
+bincode = "1.3.3"
+serde = { version = "1.0", features = ["derive"] }
+thiserror = "1.0"
