+++ crates/kotoba-types/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "kotoba-types"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+indexmap = { version = "2.0", features = ["serde"] }
+serde_json = "1.0"
+thiserror = "1.0"
